Fix 2198937: Home becomes unresponsive after picking wallpaper
THe problem was that the code was relying on mWaitingForResult being
cleared in onActivityResult. This was not being called because the
wallpaper choosing activity was not being invoked with a result code.

We may want to rethink blocking touch events when isWorkspaceLocked().
There may be other places where we can get stuck like this.