Fix bugs 226264 and 2265943 - Home stopped responding to touch, but all-apps & notification panel still work
I knew it was a problem with mWaitingForResult, which I'm still not convinced we need, but
Jeff Sharkey was able to find repro steps, so he's my new hero.  The steps were:
  1. Open add menu (long pressing or menu > add)
  2. Turn off the screen
  3. Turn back on & unlock
  4. Poof.
The problem was that when we get ACTION_CLOSE_SYSTEM_DIALOGS intents, we're dismissing the dialogs,
not canceling them.  I think we actually want to cancel them, but there's no easy way to do that
with the Activity class, so just clear mWaitingForResult by hand in closeSystemDialogs().