Last minute hack to deliver touch events to widget in landscape mode. Bug #2562729
In landscape, the left and right dots buttons are 93dip wide; this makes them overlap
with widgets at the bottom of the screen. The dots do not need to be that wide but
we chose this size to make it easier to tap them in portrait. To avoid issues in
landscape, this change introduces a new type of ImageView that can ignore touch
events in a certain zone. This was easier and cheaper than re-cutting 36+ assets.

Change-Id: Id261fba41a43dede943e72060e39e87658e4b0df