When the SD card comes back, we need to show the apps on it.
The problem was that we were doing startLoader, but once the
apps are loaded, all that does is re-bind everything.  Before
a recent change of mine, we were clearing mWorkspaceLoaded and
mAllAppsLoaded.  But that's not good either, because it meant
that there needed to be a lock that the main thread and the
worker thread were both acquiring, which was causing ANRs.

Instead, now we schedule two messages - one to add the apps
back into the list, and one to update everything to use
those new icons.

Bug: 3065061
Change-Id: If9d36d4b55a7ba72fcf48a5ed8941bb9466bebe6