When the launcher is paused and we reload stuff in the background, we need to re-re-load it in onResume.
Otherwise we can load widgets and other resources from the wrong Configuration.
This doesn't completely fix the bug, but it makes it much less likely.  We tell
the launcher once at the beginning of starting a reload because of SD cards coming
back, and once when we bind.

cherry pick of I99ee6af38bef91e261832bad4dec978a5d4a8b3d

Bug: 3126698
Change-Id: I917bdb3982e3eea4924c6e9a8f3c037fd493f415