Rocket Launcher!
This is an Android Dream, which is kind of like an
interactive screensaver implemented as an Activity.

The Rocket Launcher uses the Launcher's own icon cache to
show the user's installed apps in hyperspace. Tap the screen
to slow down to impulse power.  Then, tap an icon to launch.

(depends on I4559a95 for android.dreams in support lib)

Change-Id: Ie6ce34ec1c863667faeada4f39b225f3946e8633