Updated bug fix for Launcher reloading on mcc update.
Store the previousConfig's mcc value only instead of configuration object.
The mPreviousConfig object was a reference to the same object ResourceManager
was using, which can change by the time LauncherModel gets the intent for
ACTION_CONFIGURATION_CHANGED.
This was causing Launcher to not reload all apps as the currentConfig and previousConfig always matched.

Bug 5040470

Change-Id: Ie7102cc233177e05cb0e0e2ae4240309c16ebf66