Stop animating All Apps during transition to Home
One of the sources of jank in launcher is during the All Apps -> Home
transition. specifically, if the user has started a fling operation (where
we are animating between pages in All Apps) and then hits the Home button,
we continue the fling animation while also doing the transition to Home
scale/fade animations. This causes a lot of work for launcher, particularly because
the fling animation is causing the All Apps layer to get recreated on every frame.

The fix is to simply pause the fling animation, then snap to its end state when the
animation to Home is complete. We also need to pause/snap the scroll indicator animation,
because it's fading animation causes the same layer-recreation jank that the fling
itself causes.

Issue #7387124 Home <-> All Apps transition animation is janky while flinging

Change-Id: Icbcaf2d5b3b2f6ce8fd7419419d258248aa1475b