Trebuchet: Centers horizontally the drawer on vertical orientation
The drawer (AppsCustomizePagedView) disable the use of centers
horizontally && vertical every page. There is no problem with that on
device with fixed grid, and on devices with isLargeScreen on horizontal
orientation aligning the cellpagedview the top looks good. But when
orientation is vertical, the cellpagedview is aligned to the left.
In this case, this change forces the cellpagedview to center horizontally,
which give to the cellpagedview the same good look than horizontal orientation.

[before] https://sharecloud.ruesga.com/public.php?service=files&t=5fee1bd3e68fd28eafcaff13357956db

[after] https://sharecloud.ruesga.com/public.php?service=files&t=0beb9a41d6ad86fef089fb9f643ff51c

Change-Id: Ib5c123ce4e978500fc30cc039a62c590d75b09f7
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>