Fix for low aspect ratio (4x3) tablets.
The recent changes (http://review.cyanogenmod.org/#/c/42991/) cause
problems for tablets with a 4x3 screen such as Tenderloin. The icon
spacing is not calculated correctly. This is stated in the code for
Workspace.java at line 441:
"//TODO: This code currently fails on tablets with an aspect ratio < 1.3."
This change fixes that failure for such tablets.
Besides setting the grid size, stretching is required to get useable
screens. This reenables stretching as well as grid size setting if
the maintainer has specified
    <bool name="config_workspaceTabletGrid">true</bool>
in an overlay. There is no change to the default behavior.

Change-Id: I71b51a5ffe9911f1bb2da0b33930f73950efc782