Trebuchet : Check item placement when setting defaultWorkspace
In its current state, the launcher blindly adds any item specified
in the default workspace (even those that overlap). Consequently, it
throws an error each time it tries to load the database.

This patchset addresses that by dismissing insertion of overlapping items.
This allows for fallbacks to be specified in the default_workspace.xml

Change-Id: I08a6bd699defb908565ca2f47c855516eecf5643