Improve margins for small tablets
We are currently allocating too much space on the side margins in
landscape on small tablets with the searchbar hidden.
This fixes the logic for determining the space.
The margins to the left side and the hot seat are the same,
rather than wider on the left side as they are now.

Change-Id: Ic57aa4f850253f1ec1f8858b4a137c8d8a2ad534