Improve margins for all tablets
My last patch caused the Remove and App Info icons to be cut off
and distorted. This change increases the space so that there is
no distortion on full sized or small tablets.
A problem remains that was present before. When you have the
searchbar invisible in landscape on a small tablet, the Remove
and App Info icons are not displayed. It works fine on large tablets,
on small tablets with the search bar visible and on small
tablets in portrait.

Change-Id: I61ae8d70785878026c1207f3677d667cc49d7c50