Give GEL a way to clear the typed text when it starts searching
...otherwise it'll be re-used every time the user starts a new
query. We take care not to change the way that typed text gets
cleared (in onWindowVisibilityChanged()) in other cases.

GEL will actually use this text in ag/537329. This CL will cause
the GEL build to break until that change is submitted, so
I'll be sure to submit them both close to each other.

Bug: 16953058
Change-Id: I7ed613e69eb7a38b1a5e6108715ca614b59fd24d