Enabling accessibility dragging with the framework drag driver.
We don't want to start a framework DND (with drop shadow etc.) in the accessibility mode.

Besides, framework DND fails to start and immediately cancels the DND operation
if it's started in the accessibility mode. Presumably, this happens because
the finger doesn't press the screen when this happens.

The right solution is not to instantiate a drag driver at all when an accessible
DND starts, which is what this CL does.

Bug: 22028725
Change-Id: Ic743ba3f8de037c15d4e70e3b7687cdd3518b2a3