startAnimationToWorkspace() handles coming from another workspace state.
Previously, it assumed that the transition started from an overlay -
either all apps or the widget picker. But this isn't true, because
the transition is also used when transitioning between workspace
states such as from normal to overview. Properly handling this case is
critical for the workspace to correctly manage the transition. For
instance, it wasn't setting mIsSwitchingState to true, causing issues
such as invisible overview panels in landscape mode.

Change-Id: I9c06a345233d366669972359c58c3427a518e2b9