[DO NOT MERGE] Revert "Fix broken make target launcher_protoutil_lib Bug:26406415"
In branch ub-launcher3-burnaby,
libprotobuf-java-2.3.0-nano can be linked, because the protobuf project is tracking lmp tree

In branch ub-launcher3-burnaby-polish,
libprotobuf-java-nano can be linked, because the
protobuf project is tracking master tree.

Hence the original CL should be reverted in ub-launcher3-burnaby tree but should not be merged automatically
to downstream trees.

This reverts commit 4ae00b1b8e853f45770c833085179eb382220d78.

Change-Id: I137756b79e2c12d36408073bdd6cddc773a4f168