Make padding consistent between shortcuts and the app icon.
This change makes the padding consistent regardless of where the app
is (e.g. folder vs workspace vs all apps) by ignoring the app's
padding and adding our own to the shortcuts container.

Also note that this padding is relative to the icon, excluding the
text beneath it. So we also hide the text when the container opens
downwards, and re-show it when the container closes.

Bug: 30604007
Change-Id: I6e51c4983a8b5d495833f86e483ebaa229ed2099