Allowing Launcher to be backed up while in foreground
Full backup requies the process to be killed, and hence avoids backing
up apps which are in foreground. Since the launcher is more likely to be
the foreground app when device is locked, it never gets backuped up.

Note that full backup only runs when the screen is off, and the launcher
activity is restarted as soon as the full backup is complete (without
waiting for the screen to turn on).
There would not be much visible impact to the user.

Bug: 30970354
Change-Id: Icc16b35682c5b8d59316bd9d68774d38cbee47dd