Remove setCurrentPlayTime call as it causes a crash.
In the previous change [1], we remove the duration
from the AnimatorSet so that the preview item
animations can finish faster than the other animations.
This causes the call to setCurrentPlayTime to crash,
because we never set a duration on the AnimatorSet.

This can be solved by creating a ValueAnimatorSet (future CL).

1: If63d4d032078ff4e28b72a09e35da3a71f1d0e27

Bug: 35064148
Change-Id: Ieed21e1ba8e669132188ce15fc20195ac841d14c