Backport of Android O exclusive Features in Pixel Launcher
This includes the Google implementation of Notification Badges, the new design of shortcuts (including the notification view), and also the widget bottom view.
In theory this commit should be more or less compatible with Launcher3, probably requires a lot of manual merging though.

If anyone imports this anywhere I would be really glad to be properly credited as over 20 hours of work were put into this, and I definetly didn't do this to just be kanged!