Animate Folder elevation halfway thru transition.
The current implementation had the Folder elevation animate in
after the Folder transition was done completing. To make it appear
more smooth, we animate the Folder elevation halfway thru the
transition.

We will keep the separate FolderIcon shadow fade-in animation on
the closing transition otherwise the shadow will be apparent in the
"preview."

Bug: 62787582
Bug: 35064148
Change-Id: Ifd3d84214b0600b6651005c8e65d90b0d8ade422