Polish Folder open animation during drag and drop.
We now use the current FolderIcon background radius as the
clipping size for the Folder, which results in a smoother
hand-off between the two.

Before this change, the Folder would center itself against the FolderIcon
during the hand-off. This resulted in the preview items being fully-visible
in the FolderIcon, to being clipped in the Folder.

Bug: 63709477

Change-Id: I1da187485f28dde92d171729b011ada47a506f85