Fix tab selected color is not rendered properly
We used to use mSelectedPosition for both tab / indicator position.
However, they are not always in sync.
For example, if user taps on a tab, that tab will be in "selected state"
right away when we are still showing the sliding animation.

The fix is to use two separate variables to keep track of
tab / indicator position.

Fix: 70778492

Change-Id: I9b01fa2108b256448d3de0113f9bed75043afefd