Simplifying/fixing logic for ignoring touches in quick scrub
Now we directly set a static boolean as soon as we get the
onQuickScrubStart and end binder callbacks, instead of having
a member variable in QuickScrubController. Notably, QSC was
getting onQuickScrubStart() when doing quick switch, but not
onQuickScrubEnd(). This is a byproduct of some other unrelated
logic, which the new approach circumvents.

Bug: 72444652
Change-Id: I0a3d1ae2e9970cfffc38a25df0b8f04f26cb4cfa