Set remote animations duration based on animation being run
Previously we always set the duration to 500, the app launch duration,
but now the animation can resolve to launching a recent task instead,
which uses a shorter duration. This led to effects such as the status
bar transitioning a bit late on those transitions.

Since we don't techinically know whether we are launching an app vs a
recent task until the animation starts (since we need to check the
opening target and corresponding task id), for now we just make an
educated guess based on the view type and launched component.

Change-Id: I8ebf10d24081d474a48a1eea55419651e2214545