Solving losing [Alt+]Tab events
When you are in touch mode with Overview open, and press Tab or Alt+Tab,
it will be lost. This is because the upon pressing this key(s), app
gets switched to the keyboard mode. When this happens, the first
navigation ket gets ignored, and instead of navigation, we get focus
and highlight on the first focusable element.

The concept of the touch vs keyboard mode doesn’t fit well Overview
mode, hence the solution is to not let the recents activity to leave the
touch mode. We always keep RecentsView focused when visible. Android
honors the existing focus, and doesn't’ attempt transitioning from the
touch mode.

Bug: 73090995
Test: Manually fetch Overview, press Tab or Alt-Tab. Drag a task with
touch a bit, then press [Alt+] Tab. Repeat. Do the same with Overview
opened via Alt-Tab. Also: before, after and between, also check that Tab
keeps working on the workspace.

Change-Id: If6890ee80e08b03580051cceeac1be9a89ad3656