Focusing on the second task on Alt+Tab
Reusing the quick scrub codepath. I’m postponing the cleanup with
removing duplicated code and isolating the reused code so that it’s
separated from touch events.

I’m accumulating the cleanup notes in a doc, and will create a Q bug
from them at the end.

This CL leaves mOverviewCommandHelper.onOverviewShown/Hidden() unused,
but we may use them later when implementing the case when System UI
invokes the callback with different params.

Note that the noew code works even when Alt+Tab is pressed while the
Overview is already open.

Bug: 73090995
Test: Alt+Tab. Alt+Tab while Overview is open.
Change-Id: Iba05c5628959b3640ddd643dead0fba3ae4714c1