Show overview tiles in the orientation screenshot was taken
If the orientation of the screenshot doesn't match the thumbnail view,
we rotate the screenshot 90 degrees. We don't know whether a landscape
screenshot was seascape, so we just assume it's not (i.e. landscape
screenshots always rotate +90 degrees when shown in portrait).

Portrait screenshots rotate so that the bottom aligns with the nav bar,
e.g. rotate +90 in seascape and -90 in landscape

Currently guarded by the flag OVERVIEW_USE_SCREENSHOT_ORIENTATION.

Bug: 74552612
Change-Id: I438e45d89b54ffe41950c8bb9abdbb9a1c33aa40