Added two touch slops for easier home press (3/3)
The first slop starts the drag either for quick step or scrub. At this
point either launcher can animate swipe up or the home button moves for
scrub, however the home button is still active and can be pressed or
long pressed. The second slop is for activating quick step or scrub.
Similar to before when an operation has activated, the home button
cannot be pressed or long pressed. This allows the home button to have a
larger area to be clicked.

Change-Id: Ie792845a09ecf5d8bdbeae9e8ef7210e2299c69d
Fixes: 76430825
Test: scrub or swipe up